// src/app/models/vetCare.ts
export interface VetCare {
    uid?: string;
    name?: string;
    price?: number;
    typeservice?: string;
    imageUrl?: string;
    asociadoID?: string;
    asociadoImageUrl?: string;
    asociadoName?: string;
    asociadoHorario?: string;
    asociadoDescription?: string;
    catId?: string;
    catName?: string;
    color?: string;
    asociadoDirection?: string;
    asociadoWebsite?: string;
    asociadoPhone?: string;
    isActive?: boolean;
    amountVent?: number;
    date?: string;
}